+++ lib/src/xml/mixins/has_writer.dart
@@ -1,7 +1,8 @@
 library xml.mixins.has_writer;
 
-import '../entities/default_mapping.dart';
 import '../entities/entity_mapping.dart';
+import '../nodes/node.dart';
+import '../utils/predicate.dart';
 import '../visitors/pretty_writer.dart';
 import '../visitors/writer.dart';
 import 'has_visitor.dart';
@@ -9,38 +10,45 @@ import 'has_visitor.dart';
 /// Mixin to serialize XML to a [StringBuffer].
 mixin XmlHasWriter implements XmlHasVisitor {
   /// Return a default XML string of this object.
-  @override
-  String toString() => toXmlString();
-
   /// Return an XML string of this object.
   ///
-  /// If `pretty` is set to `true` the output is nicely reformatted, otherwise
+  /// If [pretty] is set to `true` the output is nicely reformatted, otherwise
   /// the tree is emitted verbatim.
   ///
-  /// The option `indent` is used only when pretty formatting to customize the
-  /// indention of nodes, by default nodes are indented with 2 spaces.
+  /// The [entityMapping] defines how character entities are encoded into the
+  /// resulting output.
   ///
-  /// The option `newLine` is used only when pretty formatting to customize the
-  /// printing of new lines, by default the standard new-line `'\n'` character
-  /// is used.
-  ///
-  /// The option `level` is used only when pretty formatting to customize the
-  /// initial indention level, by default this is `0`.
-  String toXmlString(
-      {bool pretty = false,
-      XmlEntityMapping entityMapping = const XmlDefaultEntityMapping.xml(),
-      int level = 0,
-      String indent = '  ',
-      String newLine = '\n'}) {
+  /// All other options are used for pretty printing only:
+  /// - The option [indent] defines the indention of nodes, by default nodes
+  ///   are indented with 2 spaces.
+  /// - The option [newLine] defines the printing of new lines, by default
+  ///   the standard new-line `'\n'` character is used.
+  /// - The option [level] customizes the initial indention level, by default
+  ///   this is `0`.
+  /// - If the predicate [preserveWhitespace] returns `true`, the whitespace
+  ///   characters within the node are preserved. By default all whitespace
+  ///   is normalized.
+  String toXmlString({
+    bool pretty = false,
+    XmlEntityMapping entityMapping,
+    int level,
+    String indent,
+    String newLine,
+    Predicate<XmlNode> preserveWhitespace,
+  }) {
     final buffer = StringBuffer();
     final writer = pretty
         ? XmlPrettyWriter(buffer,
             entityMapping: entityMapping,
             level: level,
             indent: indent,
-            newLine: newLine)
+            newLine: newLine,
+            preserveWhitespace: preserveWhitespace)
         : XmlWriter(buffer, entityMapping: entityMapping);
     writer.visit(this);
     return buffer.toString();
   }
+
+  @override
+  String toString() => toXmlString();
 }