+++ lib/src/xml/utils/predicate.dart
@@ -1,0 +1,4 @@
+library xml.utils.predicate;
+
+/// Predicate function
+typedef Predicate<T> = bool Function(T value);