+++ test/all_test.dart
@@ -15,6 +15,7 @@ import 'parse_test.dart' as parse_test;
 import 'query_test.dart' as query_test;
 import 'regression_test.dart' as regression_test;
 import 'utils_test.dart' as utils_test;
+import 'visitor_test.dart' as visitor_test;
 
 void main() {
   group('builder', builder_test.main);
@@ -30,4 +31,5 @@ void main() {
   group('query', query_test.main);
   group('regression', regression_test.main);
   group('utils', utils_test.main);
+  group('visitor', visitor_test.main);
 }