+++ CHANGELOG.md
@@ -4,11 +4,12 @@
 
 * Deprecate standalone `XmlDocument parse(String input)` method, and introduce factory methods in the respective nodes `XmlDocument.parse(String input)` and `XmlDocumentFragment.parse(String input)`.
 * Introduce getters and setters for `XmlNode.innerText` (in most cases an alias to `XmlNode.text`), `XmlNode.innerXml` and `XmlNode.outerXml`.
-* Improved support for `XmlDocumentFragment` accross the library.
+* Improved support for `XmlDocumentFragment` across the library.
 * Remove the `XmlDocument.text` override, which returned `null`.
 * Add `XmlNode.replace(XmlNode other)` to make it easier to replace nodes in an existing tree.
 * Add `XmlNode.getElement(String name)` as a shortcut to find the first child element with a given name.
 * Add `XmlNode.firstElementChild` and `XmlNode.lastElementChild` to easy access the first/last child element.
+* Add support to selectively disable  whitespace normalization while pretty-printing, for example `document.toXmlString(pretty: true, preserveWhitespace: (node) => node is XmlElement && node.name.local == 'pre')` would keep everything within `<pre>` tags as-is.
 
 ## 4.1.0